% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_goodies.R
\name{added_da}
\alias{added_da}
\title{Test added diagnostic accuracy (estimated with AUC) dued to
variable insertion in a logistic model}
\source{
idea in Knottnerus Buntinx (editors), Evidence base of
    clinical diagnosis, 2008, Wiley Blackwell, pag 157
}
\usage{
added_da(formula, data, perm_test_sequence = FALSE, plot_roc = TRUE,
  col = "black", lty = 1, legend_style = c("covariates_auc",
  "numbers_auc", "none"), test_style = c("all_combs", "BvsC", "AvsC"),
  test_param = list())
}
\arguments{
\item{formula}{a glm full formula}

\item{data}{data.frame}

\item{plot_roc}{plot the overlapping rocs of the models estimated}

\item{col}{color of the rocs}

\item{lty}{lty of the rocs}

\item{legend_style}{covariates_auc for the right hand formula and
auc, numbers_auc for number between parenthesis and auc, none
for avoiding printing legend}

\item{test_style}{all_combs for testing all the combinations, AvsC
for comparing all the models to the one with only one
covariate, BcsC to compare the models with the one identical
excepting for the last covariate added}

\item{test_param}{list of parameter to be passed to pROC::roc.test}

\item{perm_tests_sequence}{permutate model covariate sequence (checking
all possible models/sequences given the covariates specified)}
}
\description{
The function add one terms per iteration, estimate the model,
calculates the roc and finally compare them with z test using
delong variance (by default, but other tests can be asked for,
using pROC::roc.test).
}
