% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/da_compare.R
\name{bonett_price}
\alias{bonett_price}
\title{Confidence interval for the difference between two
paired proportions in a 2x2 table}
\usage{
bonett_price(n12, n21, N, alpha = 0.05)
}
\arguments{
\item{n12}{corresponding cell frequency in 2x2 setting}

\item{n21}{corresponding cell frequency in 2x2 setting}

\item{N}{table total frequency}

\item{alpha}{type I error}
}
\description{
This function computes the Wald confidence interval with
Bonett-Price Laplace adjustment confidence interval for the
difference between two paired proportions. It gives the estimate as
well.
}
\examples{

# table V fagerland etal, recommended tests and confindence intervals
# for paired binomial proportions  ,stat in med 2014

bonett_price(n12 = 1, n21 = 7, N =21)
}
\references{
Fagerland MW, Lydersen S, Laake P. Recommended tests and confidence
    intervals for paired binomial proportions.  Stat Med. 2014 Jul
    20;33(16):2850-75. doi: 10.1002/sim.6148. Epub 2014 Mar 20.
}
