% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/da.R
\name{da}
\alias{da}
\title{Calculate diagnostic accuracy measures for binary measures (test, reference
standard)}
\usage{
da(
  test = NULL,
  refstd = NULL,
  alpha = 0.05,
  digits = 4L,
  positive_first = TRUE,
  ppv_npv_prev = NULL,
  ppv_npv_force_unadj = FALSE,
  ppv_npv_ci_method = c("mercaldo", "clopper.pearson")
)
}
\arguments{
\item{test}{test variable (dichotomic factor or logical)}

\item{refstd}{reference standard (dichotomic factor or logical)}

\item{alpha}{type I error for two sided confidence interval}

\item{digits}{rounding digits}

\item{positive_first}{logical: display positive test and reference
standard in first row column}

\item{ppv_npv_prev}{prevalence adopted for ppv and npv confidence
interval (if NULL it's estimated from sample)}

\item{ppv_npv_force_unadj}{force unadjusted (standard logit) ppv
npv estimates (and confidence interval) if at least one between
NPV or PPV is 0 or 1}
}
\description{
Calculate diagnostic accuracy measures for binary measures (test,
reference standard) and confidence intervals (two sided binomial
for all the indexes except PPV and NPV, calculated via bdpv a-la
Mercaldo)
}
\examples{

## CASS Example (Pepe pag 22)
db <- dadb(tn = 327, fn = 208, fp = 115, tp = 815)
with(db, da(test = test, refstd = refstd))

## Alzheimer Example (Mercaldo 2007)
db <- dadb(tn = 288, fn = 178 , fp = 87, tp = 240)
with(db, da(test = test, refstd = refstd, ppv_npv_prev = .03))

}
