% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/da.R
\name{da_compare}
\alias{da_compare}
\title{compare diagnostic accuracy measures for 2 binary tests evaluated
in a paired setting (each test evaluated each patients) having a
reference tandard}
\usage{
da_compare(test1 = NULL, test2 = NULL, refstd = NULL, alpha = 0.05,
  test1_lab = "Test 1", test2_lab = "Test 2",
  refstd_lab = "Refence standard")
}
\arguments{
\item{test1}{oldest test available, logical value with FALSE = non
diseased and TRUE diseased or factor with two levels}

\item{test2}{new test, same coding as test1}

\item{refstd}{reference standard diagnosis same coding as test1}

\item{alpha}{Type I error}
}
\description{
compare diagnostic accuracy measures for 2 binary tests evaluated
in a paired setting (each test evaluated each patients) having a
reference tandard
}
