% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_goodies.R
\name{hosmerlem}
\alias{hosmerlem}
\title{Hosmer Lemeshow test}
\usage{
hosmerlem(y, yhat, g = 10)
}
\arguments{
\item{y}{factor with indicator variable}

\item{yhat}{predicted probability}

\item{g}{number of groups}
}
\description{
Hosmer Lemeshow test for
}

