% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{quick_roc}
\alias{quick_roc}
\title{A quick ROC plot.}
\usage{
quick_roc(
  test = NULL,
  refstd = NULL,
  roc_params = list(direction = c("auto", "<", ">")),
  plot = TRUE,
  plot_params = list(print.thres = TRUE, mar = par("mar"), print.thres.pattern =
    "Thresh: \%.2f\\nSp: \%.2f\\nSe: \%.2f", main = ""),
  plot_auc = TRUE,
  plot_auc_coords = c(0.8, 0.2),
  plot_auc_pos = 1,
  plot_auc_cex = 1
)
}
\arguments{
\item{test}{Quantitative marker}

\item{refstd}{Reference standard (binary variable)}

\item{roc_params}{list of parameters passed to pROC::roc}

\item{plot}{Logical. Plot ROC curve?}

\item{plot_params}{list of parameters passed to pROC::plot.roc}

\item{plot_auc}{add auc estimates to the plot}

\item{plot_auc_coords}{cordinates for auc estimates in the plot}

\item{plot_auc_pos}{positioning/alignment for auc estimates in the plot}

\item{plot_auc_cex}{cex for auc estimates in the plot}
}
\value{
The function plot the graph and return a list with ROC statistics
}
\description{
A quick ROC plot. It's a simple wrapper around \code{pROC::roc},
which adds Youden's criterion optimal cutoff and conditional
positioning for AUC (+CI) estimates.
}
