% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_roc.R
\name{roc_with_boxplot}
\alias{roc_with_boxplot}
\title{ROC graph with boxplot below}
\usage{
roc_with_boxplot(test, refstd, direction = c("auto", "<", ">"),
  layout_heights = c(3.5, 1))
}
\arguments{
\item{test}{Marker}

\item{refstd}{Reference standard (binary variable)}

\item{direction}{Charachter. Direction passed to \code{pROC::roc}}

\item{layout_heights}{layout heights for roc/boxplot image proportions}
}
\description{
ROC graph with boxplot below
}
