% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_goodies.R
\name{hosmerlem}
\alias{hosmerlem}
\title{Hosmer Lemeshow test}
\usage{
hosmerlem(mod, g = 10)
}
\arguments{
\item{mod}{glm (family = binomial) model}

\item{g}{number of groups}

\item{y}{factor with indicator variable}

\item{yhat}{predicted probability}
}
\description{
Hosmer Lemeshow test for logistic regression
}
