% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_goodies.R
\name{logistic_roc}
\alias{logistic_roc}
\title{Plot logistic model's ROC with AUC and confidence interval}
\usage{
logistic_roc(mod)
}
\arguments{
\item{mod}{a glm (family = binomial) model}
}
\description{
Plot logistic model's ROC with AUC and confidence interval
}
